backend/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
auto-py-to-exe==2.43.3
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Eel==0.16.0
flake8==7.0.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.6
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.2
webcolors==1.11.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
gunicorn==20.1.0